Add my_inject method
This commit adds the last enumerable method, my_inject
