+++ enumerable.rb
@@ -1,3 +1,5 @@
+require './utils'
+
 module Enumerable
   def my_each
     return to_enum unless block_given?
@@ -24,39 +26,23 @@ module Enumerable
   def my_all?(parameter = nil)
     return my_all_param(parameter) unless parameter.nil?
 
-    (my_each { |item| return false unless yield(item) }) if block_given?
-
-    my_each { |item| return false if item.nil? || item == false }
-    true
-  end
-
-  def my_all_param(parameter)
-    if parameter.class == Class
-      my_all? { |item| item.class == parameter }
-    elsif parameter.class == Regexp
-      my_all? { |item| item.match(parameter) }
+    if block_given?
+      my_each { |item| return false unless yield(item) }
     else
-      my_all? { |item| item == parameter }
+      my_each { |item| return false if item.nil? || item == false }
     end
+    true
   end
 
   def my_any?(parameter = nil)
     return my_any_param(parameter) unless parameter.nil?
 
-    (my_each { |item| return true if yield(item) }) if block_given?
-
-    my_each { |item| return true if !item.nil? && item != false }
-    false
-  end
-
-  def my_any_param(parameter)
-    if parameter.class == Class
-      my_any? { |item| item.class == parameter }
-    elsif parameter.class == Regexp
-      my_any? { |item| item.match(parameter) }
+    if block_given?
+      my_each { |item| return true if yield(item) }
     else
-      my_any? { |item| item == parameter }
+      my_each { |item| return true if !item.nil? && item != false }
     end
+    false
   end
 
   def my_none?(parameter = nil)
@@ -66,24 +52,6 @@ module Enumerable
 
     my_each { |item| return false if yield(item) }
     true
-  end
-
-  def no_block_given
-    return true if (length == 1 && self[0]) || empty?
-
-    my_each { |item| return false if item }
-
-    true
-  end
-
-  def my_none_param(parameter)
-    if parameter.class == Class
-      my_none? { |item| item.class == parameter }
-    elsif parameter.class == Regexp
-      my_none? { |item| item =~ parameter }
-    else
-      my_none? { |item| item == parameter }
-    end
   end
 
   def my_count(parameter = :no_parameter)
@@ -107,5 +75,17 @@ module Enumerable
 
     length.times { |index| newarray << yield(self[index]) }
     newarray
+  end
+
+  def my_inject(param_a = nil, param_b = nil)
+    tmp_array = to_a + []
+    tmp_array.unshift(param_a) if literal?(param_a)
+    return evaluate_with(tmp_array, param_b) if param_a && param_b
+    return evaluate_with(tmp_array, param_a) if !param_a.nil? && !literal?(param_a)
+
+    result = tmp_array.shift
+    tmp_array.my_each { |item| result = yield result, item }
+
+    result
   end
 end
